+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/TopicOperator.java
@@ -12,15 +12,30 @@ import io.fabric8.kubernetes.api.model.Volume;
 import io.fabric8.kubernetes.api.model.VolumeMount;
 import io.fabric8.kubernetes.api.model.ServiceAccount;
 import io.fabric8.kubernetes.api.model.ServiceAccountBuilder;
+import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.api.model.extensions.Deployment;
 import io.fabric8.kubernetes.api.model.extensions.DeploymentStrategy;
 import io.fabric8.kubernetes.api.model.extensions.DeploymentStrategyBuilder;
 import io.strimzi.api.kafka.model.KafkaAssembly;
 import io.strimzi.operator.cluster.operator.resource.RoleBindingOperator;
+import io.strimzi.api.kafka.model.Resources;
+import io.strimzi.api.kafka.model.Sidecar;
+import io.strimzi.certs.CertAndKey;
+import io.strimzi.certs.CertManager;
+import io.strimzi.certs.Subject;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
 import java.util.ArrayList;
+import java.util.Base64;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+
+import static java.util.Collections.singletonList;
 
 /**
  * Represents the topic operator deployment
@@ -32,7 +47,12 @@ public class TopicOperator extends AbstractModel {
      */
     public static final String TOPIC_CM_KIND = "topic";
 
+    protected static final String TOPIC_OPERATOR_NAME = "topic-operator";
     private static final String NAME_SUFFIX = "-topic-operator";
+    private static final String CERTS_SUFFIX = NAME_SUFFIX + "-certs";
+    protected static final String TLS_SIDECAR_NAME = "tls-sidecar";
+    protected static final String TLS_SIDECAR_VOLUME_NAME = "tls-sidecar-certs";
+    protected static final String TLS_SIDECAR_VOLUME_MOUNT = "/etc/tls-sidecar/certs/";
 
     protected static final String METRICS_AND_LOG_CONFIG_SUFFIX = NAME_SUFFIX + "-config";
 
@@ -51,6 +71,7 @@ public class TopicOperator extends AbstractModel {
     public static final String ENV_VAR_FULL_RECONCILIATION_INTERVAL_MS = "STRIMZI_FULL_RECONCILIATION_INTERVAL_MS";
     public static final String ENV_VAR_ZOOKEEPER_SESSION_TIMEOUT_MS = "STRIMZI_ZOOKEEPER_SESSION_TIMEOUT_MS";
     public static final String ENV_VAR_TOPIC_METADATA_MAX_ATTEMPTS = "STRIMZI_TOPIC_METADATA_MAX_ATTEMPTS";
+    public static final String ENV_VAR_TLS_ENABLED = "STRIMZI_TLS_ENABLED";
     public static final String TO_CLUSTER_ROLE_NAME = "strimzi-topic-operator";
     public static final String TO_ROLE_BINDING_NAME = "strimzi-topic-operator-role-binding";
 
@@ -63,6 +84,13 @@ public class TopicOperator extends AbstractModel {
     private int zookeeperSessionTimeoutMs;
     private String topicConfigMapLabels;
     private int topicMetadataMaxAttempts;
+
+    private Sidecar tlsSidecar;
+
+    /**
+     * Private key and certificate for encrypting communication with Zookeeper and Kafka
+     */
+    private CertAndKey cert;
 
     /**
      * @param namespace Kubernetes/OpenShift namespace where cluster resources are going to be created
@@ -166,7 +194,7 @@ public class TopicOperator extends AbstractModel {
     }
 
     protected static String defaultBootstrapServers(String cluster) {
-        return KafkaCluster.serviceName(cluster) + ":" + io.strimzi.api.kafka.model.TopicOperator.DEFAULT_BOOTSTRAP_SERVERS_PORT;
+        return KafkaCluster.headlessServiceName(cluster) + ":" + io.strimzi.api.kafka.model.TopicOperator.DEFAULT_BOOTSTRAP_SERVERS_PORT;
     }
 
     protected static String defaultTopicConfigMapLabels(String cluster) {
@@ -175,13 +203,19 @@ public class TopicOperator extends AbstractModel {
                 Labels.STRIMZI_KIND_LABEL, TopicOperator.TOPIC_CM_KIND);
     }
 
+    public static String secretName(String cluster) {
+        return cluster + io.strimzi.operator.cluster.model.TopicOperator.CERTS_SUFFIX;
+    }
+
     /**
      * Create a Topic Operator from given desired resource
      *
+     * @param certManager Certificate manager for certificates generation
      * @param resource desired resource with cluster configuration containing the topic operator one
+     * @param secrets Secrets containing already generated certificates
      * @return Topic Operator instance, null if not configured in the ConfigMap
      */
-    public static TopicOperator fromCrd(KafkaAssembly resource) {
+    public static TopicOperator fromCrd(CertManager certManager, KafkaAssembly resource, List<Secret> secrets) {
         TopicOperator result;
         if (resource.getSpec().getTopicOperator() != null) {
             String namespace = resource.getMetadata().getNamespace();
@@ -198,10 +232,63 @@ public class TopicOperator extends AbstractModel {
             result.setLogging(tcConfig.getLogging());
             result.setResources(tcConfig.getResources());
             result.setUserAffinity(tcConfig.getAffinity());
+            result.generateCertificates(certManager, secrets);
+            result.setTlsSidecar(tcConfig.getTlsSidecar());
         } else {
             result = null;
         }
         return result;
+    }
+
+    /**
+     * Manage certificates generation based on those already present in the Secrets
+     *
+     * @param certManager CertManager instance for handling certificates creation
+     * @param secrets The Secrets storing certificates
+     */
+    public void generateCertificates(CertManager certManager, List<Secret> secrets) {
+        log.debug("Generating certificates");
+
+        try {
+            Optional<Secret> clusterCAsecret = secrets.stream().filter(s -> s.getMetadata().getName().equals(getClusterCaName(cluster)))
+                    .findFirst();
+            if (clusterCAsecret.isPresent()) {
+                // get the generated CA private key + self-signed certificate for each broker
+                clusterCA = new CertAndKey(
+                        decodeFromSecret(clusterCAsecret.get(), "cluster-ca.key"),
+                        decodeFromSecret(clusterCAsecret.get(), "cluster-ca.crt"));
+
+                Optional<Secret> topicOperatorSecret = secrets.stream().filter(s -> s.getMetadata().getName().equals(TopicOperator.secretName(cluster))).findFirst();
+                if (!topicOperatorSecret.isPresent()) {
+                    log.debug("Topic Operator certificate to generate");
+
+                    File csrFile = File.createTempFile("tls", "csr");
+                    File keyFile = File.createTempFile("tls", "key");
+                    File certFile = File.createTempFile("tls", "cert");
+
+                    Subject sbj = new Subject();
+                    sbj.setOrganizationName("io.strimzi");
+                    sbj.setCommonName(TopicOperator.topicOperatorName(cluster));
+
+                    certManager.generateCsr(keyFile, csrFile, sbj);
+                    certManager.generateCert(csrFile, clusterCA.key(), clusterCA.cert(), certFile, CERTS_EXPIRATION_DAYS);
+
+                    cert = new CertAndKey(Files.readAllBytes(keyFile.toPath()), Files.readAllBytes(certFile.toPath()));
+                } else {
+                    log.debug("Topic Operator certificate already exists");
+                    cert = new CertAndKey(
+                            decodeFromSecret(topicOperatorSecret.get(), "topic-operator.key"),
+                            decodeFromSecret(topicOperatorSecret.get(), "topic-operator.crt"));
+                }
+            } else {
+                throw new NoCertificateSecretException("The cluster CA certificate Secret is missing");
+            }
+
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+        log.debug("End generating certificates");
     }
 
     public Deployment generateDeployment() {
@@ -224,17 +311,31 @@ public class TopicOperator extends AbstractModel {
     protected List<Container> getContainers() {
         List<Container> containers = new ArrayList<>();
         Container container = new ContainerBuilder()
-                .withName(name)
+                .withName(TOPIC_OPERATOR_NAME)
                 .withImage(getImage())
                 .withEnv(getEnvVars())
-                .withPorts(Collections.singletonList(createContainerPort(HEALTHCHECK_PORT_NAME, HEALTHCHECK_PORT, "TCP")))
+                .withPorts(singletonList(createContainerPort(HEALTHCHECK_PORT_NAME, HEALTHCHECK_PORT, "TCP")))
                 .withLivenessProbe(createHttpProbe(livenessPath + "healthy", HEALTHCHECK_PORT_NAME, livenessInitialDelay, livenessTimeout))
                 .withReadinessProbe(createHttpProbe(readinessPath + "ready", HEALTHCHECK_PORT_NAME, readinessInitialDelay, readinessTimeout))
-                .withVolumeMounts(getVolumeMounts())
                 .withResources(resources(getResources()))
+                .withVolumeMounts(getVolumeMounts())
+                .build();
+
+        String tlsSidecarImage = (tlsSidecar != null && tlsSidecar.getImage() != null) ?
+                tlsSidecar.getImage() : io.strimzi.api.kafka.model.TopicOperator.DEFAULT_TLS_SIDECAR_IMAGE;
+
+        Resources tlsSidecarResources = (tlsSidecar != null) ? tlsSidecar.getResources() : null;
+
+        Container tlsSidecarContainer = new ContainerBuilder()
+                .withName(TLS_SIDECAR_NAME)
+                .withImage(tlsSidecarImage)
+                .withResources(resources(tlsSidecarResources))
+                .withEnv(singletonList(buildEnvVar(ENV_VAR_ZOOKEEPER_CONNECT, zookeeperConnect)))
+                .withVolumeMounts(createVolumeMount(TLS_SIDECAR_VOLUME_NAME, TLS_SIDECAR_VOLUME_MOUNT))
                 .build();
 
         containers.add(container);
+        containers.add(tlsSidecarContainer);
 
         return containers;
     }
@@ -244,11 +345,12 @@ public class TopicOperator extends AbstractModel {
         List<EnvVar> varList = new ArrayList<>();
         varList.add(buildEnvVar(ENV_VAR_CONFIGMAP_LABELS, topicConfigMapLabels));
         varList.add(buildEnvVar(ENV_VAR_KAFKA_BOOTSTRAP_SERVERS, kafkaBootstrapServers));
-        varList.add(buildEnvVar(ENV_VAR_ZOOKEEPER_CONNECT, zookeeperConnect));
+        varList.add(buildEnvVar(ENV_VAR_ZOOKEEPER_CONNECT, String.format("%s:%d", "localhost", io.strimzi.api.kafka.model.TopicOperator.DEFAULT_ZOOKEEPER_PORT)));
         varList.add(buildEnvVar(ENV_VAR_WATCHED_NAMESPACE, watchedNamespace));
         varList.add(buildEnvVar(ENV_VAR_FULL_RECONCILIATION_INTERVAL_MS, Integer.toString(reconciliationIntervalMs)));
         varList.add(buildEnvVar(ENV_VAR_ZOOKEEPER_SESSION_TIMEOUT_MS, Integer.toString(zookeeperSessionTimeoutMs)));
         varList.add(buildEnvVar(ENV_VAR_TOPIC_METADATA_MAX_ATTEMPTS, String.valueOf(topicMetadataMaxAttempts)));
+        varList.add(buildEnvVar(ENV_VAR_TLS_ENABLED, Boolean.toString(true)));
 
         return varList;
     }
@@ -289,10 +391,33 @@ public class TopicOperator extends AbstractModel {
     }
 
     private List<Volume> getVolumes() {
-        return Collections.singletonList(createConfigMapVolume(logAndMetricsConfigVolumeName, ancillaryConfigName));
+        List<Volume> volumeList = new ArrayList<>();
+        volumeList.add(createConfigMapVolume(logAndMetricsConfigVolumeName, ancillaryConfigName));
+        volumeList.add(createSecretVolume(TLS_SIDECAR_VOLUME_NAME, TopicOperator.secretName(cluster)));
+        return volumeList;
     }
 
     private List<VolumeMount> getVolumeMounts() {
-        return Collections.singletonList(createVolumeMount(logAndMetricsConfigVolumeName, logAndMetricsConfigMountPath));
+        List<VolumeMount> volumeMountList = new ArrayList<>();
+        volumeMountList.add(createVolumeMount(logAndMetricsConfigVolumeName, logAndMetricsConfigMountPath));
+        volumeMountList.add(createVolumeMount(TLS_SIDECAR_VOLUME_NAME, TLS_SIDECAR_VOLUME_MOUNT));
+        return volumeMountList;
+    }
+
+    /**
+     * Generate the Secret containing CA self-signed certificates for internal communication
+     * It also contains the private key-certificate (signed by internal CA) for communicating with Zookeeper and Kafka
+     * @return The generated Secret
+     */
+    public Secret generateSecret() {
+        Map<String, String> data = new HashMap<>();
+        data.put("cluster-ca.crt", Base64.getEncoder().encodeToString(clusterCA.cert()));
+        data.put("topic-operator.key", Base64.getEncoder().encodeToString(cert.key()));
+        data.put("topic-operator.crt", Base64.getEncoder().encodeToString(cert.cert()));
+        return createSecret(TopicOperator.secretName(cluster), data);
+    }
+
+    protected void setTlsSidecar(Sidecar tlsSidecar) {
+        this.tlsSidecar = tlsSidecar;
     }
 }